+++ CHANGELOG.rst
@@ -3,7 +3,19 @@
 Changelog
 =========
 
+Version 0.2.0 - Unreleased
+**************************
+
+* Removed every useless code;
+* Added urls, views, templatetags and templates;
+* Added some Bootstrap components for Django Cotton;
+* Added a proper layout to start;
+* Added some very basic tests;
+
+
 Version 0.1.0 - Unreleased
 **************************
 
-* First commit.
+* Generated project from template
+  `cookiecutter-sveetch-djangoapp <https://cookiecutter-sveetch-djangoapp.readthedocs.io/>`_
+  version "0.8.0";